Fixed missing tags on restart.
When we restarted the script, the previously running instance is still
alive for a moment, after the new script has initialized its tags.  When
the old one dies, it destroyes all of the widgets.  Now, we check to see
if we should delete widgets before we do.

+++ wmii.py
@@ -17,6 +17,8 @@ from pyxp.client import RPCError
 
 client = Client(namespace='wmii')
 
+_quit = False
+
 class Wmii(object):
     '''Main wmii configuration object.
 
@@ -32,8 +34,6 @@ class Wmii(object):
         self.ctl = Ctl('/ctl')
         self.tag = Tags()
         self.client = Clients()
-
-        self._quit = False
 
         self._widgets = {}
 
@@ -122,14 +122,15 @@ class Wmii(object):
 
     def restart(self):
         '''Restart configuration'''
+        global _quit
         self.execute(os.path.abspath(sys.argv[0]))
-        self._quit = True
+        _quit = True
         sys.exit()
 
     def quit(self):
         '''Quits wmii.'''
         self.ctl.write('quit')
-        self._quit = True
+        _quit = True
         sys.exit(100)
 
     def start(self, timeout=1):
@@ -145,7 +146,7 @@ class Wmii(object):
 
         # Monitor widgets
         while True:
-            if self._quit:
+            if _quit:
                 sys.exit()
             for widget in self._widgets.values():
                 try:
@@ -336,4 +337,7 @@ class Widget(object):
     label = property(get_label, set_label)
 
     def __del__(self):
+        global _quit
+        if _quit:
+            return
         self.file.remove()
